+++ packages/ui/src/contexts/ClerkUIComponentsContext.tsx
@@ -142,7 +142,6 @@ export function ComponentContextProvider({
             onAllow: p.onAllow,
             onDeny: p.onDeny,
             appearance: p.appearance,
-            enableOrgSelection: (p as any).__internal_enableOrgSelection === true,
           }}
         >
           {children}