+++ packages/ui/src/elements/contexts/index.tsx
@@ -133,7 +133,13 @@ export type FlowMetadata = {
     | 'chooseWallet'
     | 'enterpriseConnections'
     | 'organizationCreationDisabled'
-    | 'methodSelectionMFA';
+    | 'methodSelectionMFA'
+    | 'provideEmail'
+    | 'verifyDomain'
+    | 'configureCreateApp'
+    | 'configureMapAttributes'
+    | 'test-sso'
+    | 'sso-confirmation';
 };
 
 const [FlowMetadataCtx, useFlowMetadata] = createContextAndHook<FlowMetadata>('FlowMetadata');