+++ packages/nuxt/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @clerk/nuxt
 
+## 2.4.1
+
+### Patch Changes
+
+- Updated dependencies [[`7a5892f`](https://github.com/clerk/javascript/commit/7a5892f9bcaa1a6212e6e6d3741160929ffd027e)]:
+  - @clerk/backend@3.4.6
+  - @clerk/shared@4.10.1
+  - @clerk/vue@2.2.1
+
 ## 2.4.0
 
 ### Minor Changes